V1.3.0
See diff-with-0.7.5 folder for changes since V0.7.5
New UI module introduced with PyQt UI files

+++ carnatic/cdeeplearn.py
@@ -4,58 +4,68 @@ import numpy as np
 import pandas as pd
 from keras.models import Sequential
 from keras.layers import LSTM, Dropout, TimeDistributed, Dense, Activation, Embedding
-from carnatic import cparser
+from carnatic import cparser,settings
 
 #charIndex_json = "char_to_index.json"
-model_weights_directory = 'model_weights/'
+model_weights_directory = settings._APP_PATH + '/model_weights/'
 BATCH_SIZE = 16
-SEQ_LENGTH = 64
+CHORD_LENGTH = 3
 EPOCHS = 90
 MODEL_WEIGHTS_FILE = ''
 JSON_FILE = ''
 def set_parameters(batch_size=None, number_of_epochs=None,model_weights_folder= None,
-                   json_file=None, model_weights_file=None):
+                   json_file=None, model_weights_file=None,chord_length=None):
     """
     Set parameters of deep learning method:
     @param batch_size: Batch size for training. Default=16
     @param number_of_epochs: Default 90
     @param model_weights_folder: Default: "model_weights/"
-    @param json_file: File containing dictionary of unique notes. Default:"<raagam_name>_corpus|lessons.json"
-    @param model_weights_file: File containing trained weights. Default:"<raagam_name>_corpus|lessons.h5"
+    @param json_file: File containing dictionary of unique notes. Default:"<raagam_name>_lessons.json"
+    @param model_weights_file: File containing trained weights. Default:"<raagam_name>_corpus.h5"
     """
-    global BATCH_SIZE, EPOCHS, model_weights_directory, JSON_FILE, MODEL_WEIGHTS_FILE
+    global BATCH_SIZE, EPOCHS, model_weights_directory, JSON_FILE, MODEL_WEIGHTS_FILE,CHORD_LENGTH
+    if chord_length:
+        CHORD_LENGTH = chord_length
+        print('CHORD_LENGTH set to:',CHORD_LENGTH)
     if batch_size:
         BATCH_SIZE = batch_size
+        print('BATCH_SIZE set to:',BATCH_SIZE)
     if number_of_epochs:
         EPOCHS = number_of_epochs
+        print('EPOCHS set to:',EPOCHS)
     if model_weights_folder:
         model_weights_directory = model_weights_folder
+        print('model_weights_directory set to:',model_weights_directory)
     if json_file:
         JSON_FILE = json_file
-        print('JSON_FILE=',JSON_FILE)
+        print('JSON_FILE set to:',JSON_FILE)
     if model_weights_file:
         MODEL_WEIGHTS_FILE = model_weights_file
-        print('MODEL_WEIGHTS_FILE',MODEL_WEIGHTS_FILE)
+        print('MODEL_WEIGHTS_FILE set to:',MODEL_WEIGHTS_FILE)
+def are_deeplearning_parameters_defined():
+    parameters_are_set = (CHORD_LENGTH >0 ) and (BATCH_SIZE>0) and (EPOCHS>0) and (JSON_FILE != '') and (MODEL_WEIGHTS_FILE != '')
+    print("parameters_are_set to:",parameters_are_set)
+    return parameters_are_set
 def _read_batches(all_chars, unique_chars):
     length = all_chars.shape[0]
     batch_chars = int(length / BATCH_SIZE) #155222/16 = 9701
     
-    for start in range(0, batch_chars - SEQ_LENGTH, 64):  #(0, 9637, 64)  #it denotes number of batches. It runs everytime when
+    for start in range(0, batch_chars - CHORD_LENGTH, CHORD_LENGTH):  #(0, 9637, 64)  #it denotes number of batches. It runs everytime when
         #new batch is created. We have a total of 151 batches.
-        X = np.zeros((BATCH_SIZE, SEQ_LENGTH))    #(16, 64)
-        Y = np.zeros((BATCH_SIZE, SEQ_LENGTH, unique_chars))   #(16, 64, 87)
-        for batch_index in range(0, 16):  #it denotes each row in a batch.  
-            for i in range(0, 64):  #it denotes each column in a batch. Each column represents each character means 
+        X = np.zeros((BATCH_SIZE, CHORD_LENGTH))    #(16, 64)
+        Y = np.zeros((BATCH_SIZE, CHORD_LENGTH, unique_chars))   #(16, 64, 87)
+        for batch_index in range(0, BATCH_SIZE):  #it denotes each row in a batch.  
+            for i in range(0, CHORD_LENGTH):  #it denotes each column in a batch. Each column represents each character means 
                 #each time-step character in a sequence.
                 X[batch_index, i] = all_chars[batch_index * batch_chars + start + i]
                 Y[batch_index, i, all_chars[batch_index * batch_chars + start + i + 1]] = 1 #here we have added '1' because the
                 #correct label will be the next character in the sequence. So, the next character will be denoted by
                 #all_chars[batch_index * batch_chars + start + i] + 1. 
         yield X, Y
-def _get_model(batch_size, seq_length, unique_chars):
+def _get_model(batch_size, sequence_length, unique_chars):
     model = Sequential()
     
-    model.add(Embedding(input_dim = unique_chars, output_dim = 512, batch_input_shape = (batch_size, seq_length), name = "embd_1")) 
+    model.add(Embedding(input_dim = unique_chars, output_dim = 512, batch_input_shape = (batch_size, sequence_length), name = "embd_1")) 
     
     model.add(LSTM(256, return_sequences = True, stateful = True, name = "lstm_first"))
     model.add(Dropout(0.2, name = "drp_1"))
@@ -73,16 +83,22 @@ def _get_model(batch_size, seq_length, unique_chars):
     
     return model
 def _train_the_model(data, epochs):
-    global MODEL_WEIGHTS_FILE, JSON_FILE
-    char_to_index = {ch: i for (i, ch) in enumerate(sorted(list(set(data))))}
-    print("Number of unique characters in our whole tunes database = {}".format(len(char_to_index)),"\n",char_to_index) #87
-    with open(os.path.join(model_weights_directory, JSON_FILE), mode = "w") as f:
-        json.dump(char_to_index, f)
+    global MODEL_WEIGHTS_FILE, JSON_FILE,model_weights_directory
+    if os.path.exists(os.path.join(model_weights_directory,JSON_FILE)):
+        with open(os.path.join(model_weights_directory, JSON_FILE)) as f:
+            char_to_index = json.load(f)
+    else:    
+        char_to_index = {ch: i for (i, ch) in enumerate(sorted(list(set(data))))}
+        json_file = os.path.join(model_weights_directory, JSON_FILE)
+        print('json_file',json_file)
+        with open(json_file, mode = "w") as f:
+            json.dump(char_to_index, f)
         
+    print("Number of unique characters in our whole tunes database = {}".format(len(char_to_index)),"\n",char_to_index) #87
     index_to_char = {i: ch for (ch, i) in char_to_index.items()}
     unique_chars = len(char_to_index)
     
-    model = _get_model(BATCH_SIZE, SEQ_LENGTH, unique_chars)
+    model = _get_model(BATCH_SIZE, CHORD_LENGTH, unique_chars)
     model.summary()
     model.compile(loss = "categorical_crossentropy", optimizer = "adam", metrics = ["accuracy"])
     
@@ -175,9 +191,11 @@ def generate_notes_from_corpus(corpus_files:list,starting_note:str, ending_note:
     @param save_to_file: File name to which generated notes are to be written. Default=None
     @param perform_training: True/False. Default = False. 
         If True, training weights are generated even if model weight file is found  
+    @param chord_length: Number of adjacent notes to be used for prediction. Same as width for markov model
     """
     global MODEL_WEIGHTS_FILE,JSON_FILE
     print('generate_notes_from_corpus','JSON_FILE',JSON_FILE,'MODEL_WEIGHTS_FILE',MODEL_WEIGHTS_FILE)
+    print("Batch Size = {} Chord length = {} Epochs = {}".format(BATCH_SIZE,CHORD_LENGTH,EPOCHS))
     model_file = model_weights_directory + MODEL_WEIGHTS_FILE
     if not os.path.isfile(model_file) or perform_training:
         data = []
